// src/app/models/medicalCheck.model.ts
export class MedicalCheckModel{
    id?: string;
    FechaAtencion?: string;
    Temperatura?: number;
    Peso?: number;
    FrecuenciaRespiratoria?: number;
    FrecuenciaCardiaca?: number;
    EstadoAnimo?: string;
    Recomendaciones?: string;
    Medicamentos?: string;
    solicitudVisitaDomiciliariaId?: string;
}